feat : 자동차 경주 1. 전체 소스의 경로를 racing 아래 두도록 경로 수정.  - step2과제물과 혼용을 방지하고 싶었습니다. 2. 전체사용중인 메세지성 내용들을 별도의 클래스로 관리하도록 수정.
