+++ src/main/java/study/racing/common/Message.java
@@ -1,0 +1,23 @@
+package study.racing.common;
+
+public class Message {
+    /*
+    * -----------------------
+    * Common
+    * start|
+    * -----------------------
+    * RacingController
+    * 자동차 대수는 몇대 인가요?
+    * 시도할 회수는 몇 회 인가요?
+    *------------------------
+    *------------------------
+    *Validation
+    *input Value is null or blank
+    * -----------------------
+    */
+    public static final String MSG_RESULT_FORM = "start|";
+    public static final String MSG_ASK_CARS_COUNT = "자동차 대수는 몇대 인가요?";
+    public static final String MSG_ASK_ROUND_COUNT = "시도할 회수는 몇 회 인가요?";
+    public static final String MSG_ERROR_VALUE_NULL_BLANK = "input Value is null or blank";
+
+}
