+++ src/main/java/study/racing/domain/Cars.java
@@ -1,4 +1,4 @@
-package study.domain;
+package study.racing.domain;
 
 import java.util.ArrayList;
 import java.util.List;
