+++ src/test/java/study/racing/domain/CarTest.java
@@ -1,4 +1,4 @@
-package study.domain;
+package study.racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -6,7 +6,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static study.common.Common.toInt;
+import static study.racing.common.Common.toInt;
 
 class CarTest {
