+++ src/test/java/study/racing/domain/CarsTest.java
@@ -1,4 +1,4 @@
-package study.domain;
+package study.racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -6,7 +6,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static study.common.Common.toInt;
+import static study.racing.common.Common.toInt;
 
 class CarsTest {
     @ParameterizedTest
